increase size of types before multiplying and assigning

+++ Tools/CmdLine/IccProfileVisualize/iccProfileVisualize.cpp
@@ -599,9 +599,7 @@ void output3DLUT(CIccProfile *pIcc, CIccTag *tag, const std::string &sigDesc,
   icTagTypeSignature typeSig = tag->GetType();
   switch(typeSig) {
 
-// CIccTagLut8, CIccTagLut16, CIccTagLutAtoB, icSigLutBtoAType are subclases of CIccMBB
-// this might be usable for all types
-
+  // these are all subclases of CIccMBB, and can share most of the code
   case icSigLut8Type:   // CIccTagLut8
   case icSigLut16Type:  // CIccTagLut16
   case icSigLutAtoBType:  // CIccTagLutAtoB
@@ -796,17 +794,17 @@ void output3DLUT(CIccProfile *pIcc, CIccTag *tag, const std::string &sigDesc,
     }
 
 #else
-      size_t n001 = tileWidth * tileHeight * outputChannels;
-      size_t n010 = tileWidth * outputChannels;
-      size_t n100 = outputChannels;
+      size_t n001 = (size_t)tileWidth * (size_t)tileHeight * (size_t)outputChannels;
+      size_t n010 = (size_t)tileWidth * (size_t)outputChannels;
+      size_t n100 = (size_t)outputChannels;
 
       if (inputChannels < 2)
         std::swap(n010,n100);
 
-      size_t outTileStepV = imageWidth * tileHeight * outputChannels;
-      size_t outTileStepH = tileWidth * outputChannels;
-      size_t outColStep = outputChannels;
-      size_t outRowStep = imageWidth * outputChannels;
+      size_t outTileStepV = (size_t)imageWidth * (size_t)tileHeight * (size_t)outputChannels;
+      size_t outTileStepH = (size_t)tileWidth * (size_t)outputChannels;
+      size_t outColStep = (size_t)outputChannels;
+      size_t outRowStep = (size_t)imageWidth * (size_t)outputChannels;
 
       for (int z = 0; z < tiles; ++z) {
         int z2 = z % tilesWide; // tile # horiz
